requirement.txt
fastapi>=0.110.0
uvicorn>=0.27.0

python-dotenv>=1.0.1

python==3.12.7

langchain==1.0.3
langchain-huggingface>=0.1.0
langchain-openai>=0.2.0

neo4j==5.28.2

transformers>=4.35.0
sentence-transformers==2.6.1
sentencepiece>=0.1.99

datasets==4.0.0

pandas==2.2.3
torch>=2.0.0

huggingface_hub>=0.23.0,<1.0
